refactor(terminal): split the pty command module into submodules
